Split up the large SSB module
Thx @sk1p for the suggestion!
